% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_pull.R
\name{cohort_pull}
\alias{cohort_pull}
\title{Extract a cohort data frame}
\usage{
cohort_pull(.data, ...)
}
\arguments{
\item{.data}{A \code{ggconsort_cohort} object}

\item{...}{The cohort to pull as an unquoted expression}
}
\value{
A tibble for a single cohort
}
\description{
Pull a single tibble from a \code{ggconsort_cohort} object,
often for downstream analysis or deeper inspection.
}
\examples{
cohorts <- trial_data \%>\%
  cohort_start("Assessed for eligibility") \%>\%
    cohort_define(
      consented = .full \%>\% dplyr::filter(declined != 1),
      consented_chemonaive = consented \%>\% dplyr::filter(prior_chemo != 1)
    )

cohorts \%>\% cohort_pull(consented_chemonaive)
}
