% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_select.R
\name{cohort_select}
\alias{cohort_select}
\title{Subset ggconsort cohort objects}
\usage{
cohort_select(.data, ...)
}
\arguments{
\item{.data}{A \code{ggconsort_cohort} object}

\item{...}{The cohort(s) to pull as an unquoted or
\code{tidyselect} style expression}
}
\value{
A \code{ggconsort_cohort} object
}
\description{
Select a subset of cohorts from a \code{ggconsort_cohort}
object
}
\examples{
cohorts <-
  trial_data \%>\%
  cohort_start("Assessed for eligibility") \%>\%
  cohort_define(
    consented = .full \%>\% dplyr::filter(declined != 1),
    treatment_a = consented \%>\% dplyr::filter(treatment == "Drug A"),
    treatment_b = consented \%>\% dplyr::filter(treatment == "Drug B")
  ) \%>\%
  cohort_label(
    consented = "Consented",
    treatment_a = "Allocated to arm A",
    treatment_b = "Allocated to arm B"
  )

cohorts \%>\% cohort_select(consented)

cohorts \%>\%
  cohort_select(starts_with("treatment_"))

}
