% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_start.R
\name{cohort_start}
\alias{cohort_start}
\title{Initialize a new ggconsort cohort}
\usage{
cohort_start(.data, label = NULL)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{label}{A character string to describe the set of cohorts}
}
\value{
Returns a \code{ggconsort_cohort} object with
\code{$data} and \code{$labels} items
}
\description{
Creates an object with an optional label which stores
the originating source data for downstream ggconsort cohorts
}
