% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consort_start.R
\name{consort_start}
\alias{consort_start}
\title{Initialize a ggconsort object for plotting. Used internally.}
\usage{
consort_start(.data)
}
\arguments{
\item{.data}{A \code{ggconsort_cohort} object}
}
\description{
Initialize a ggconsort object for plotting. Used internally.
}
\keyword{Internal}
