% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggconsort-package.R
\docType{package}
\name{ggconsort-package}
\alias{ggconsort}
\alias{ggconsort-package}
\title{ggconsort: Creates CONSORT Diagrams for RCTs}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Takes the data from a standard status table, 
    which has a row for each patient screened and each 
    participant enrolled, and columns for each change in status,
    and uses this table to build a CONSORT diagram in ggplot. 
    CONSORT diagrams improve the transparency of reporting of
    randomized, controlled clinical trials.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/higgi13425/ggconsort}
  \item Report bugs at \url{https://github.com/higgi13425/ggconsort/issues}
}

}
\author{
\strong{Maintainer}: Peter Higgins \email{higgi13425@yahoo.com} (\href{https://orcid.org/0000-0003-1602-4341}{ORCID})

}
\keyword{internal}
