% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_count.R
\name{cohort_count}
\alias{cohort_count}
\alias{cohort_count_int}
\title{Count the number of rows in each ggconsort cohort}
\usage{
cohort_count(.data, ...)

cohort_count_int(.data, ...)
}
\arguments{
\item{.data}{A \code{ggconsort_cohort} object}

\item{...}{Cohorts to include in the output, can be quoted or unquoted
cohort names, or \pkg{tidyselect} helpers such as
[tidyselect::starts_with()].}
}
\value{
A \code{tibble} with cohort name, row number total, and label.
}
\description{
Count the number of rows in each ggconsort cohort
}
\section{Functions}{
\itemize{
\item \code{cohort_count}: Returns a tibble with cohort name, row number total
and label.

\item \code{cohort_count_int}: Returns a named vector with cohort counts.
}}

