% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_define.R
\name{cohort_define}
\alias{cohort_define}
\title{Define ggconsort cohorts}
\usage{
cohort_define(.data, ...)
}
\arguments{
\item{.data}{A \code{ggconsort_cohort} object}

\item{...}{A series of named expressions which define the cohorts}
}
\value{
The modified \code{ggconsort_cohort} object which now includes
additional \code{$data} items according to provided cohort definitions
}
\description{
Following a call to \code{cohort_start}, use \code{cohort_define}
to construct cohorts from the full source data which are appended
to the \code{ggconsort_cohort} object.
}
