% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_label.R
\name{cohort_label}
\alias{cohort_label}
\title{Add labels to ggconsort cohorts}
\usage{
cohort_label(.data, ...)
}
\arguments{
\item{.data}{A \code{ggconsort_cohort} object}

\item{...}{A series of named expressions which provide labels
corresponding to named cohorts in the \code{ggconsort_cohort} object}
}
\value{
The modified \code{ggconsort_cohort} object which now includes
additional \code{$labels} items according to provided label definitions
}
\description{
Add labels to ggconsort cohorts
}
