% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_consort.R, R/geom_consort_arrow.R,
%   R/geom_consort_box.R, R/geom_consort_line.R
\name{geom_consort}
\alias{geom_consort}
\alias{geom_consort_arrow}
\alias{geom_consort_box}
\alias{geom_consort_line}
\title{CONSORT diagram layers for ggplot2}
\usage{
geom_consort(...)

geom_consort_arrow(...)

geom_consort_box(label_color = "black", label_size = 11, label_height = 1, ...)

geom_consort_line(x, xend, y, yend, ...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggtext:geom_richtext]{ggtext::geom_richtext()}} by
\code{geom_consort_box()}, e.g. \code{fill}. Ignored by \code{geom_consort_arrow()}.}

\item{label_color}{Color of the box label text.}

\item{label_size}{Size of the box label text, in points.}

\item{label_height}{Line height of the box label text.}

\item{x, xend, y, yend}{Coordinates of the line segment drawn by
\code{geom_consort_line()}.}
}
\value{
A ggplot2 layer or list of layers that can be added to a plot.
}
\description{
\code{geom_consort()} draws the boxes, arrows, and lines of a \code{ggconsort} object
created with \code{\link[=consort_box_add]{consort_box_add()}}, \code{\link[=consort_arrow_add]{consort_arrow_add()}}, and
\code{\link[=consort_line_add]{consort_line_add()}}. It is the layer you will typically add to
\verb{ggplot(<ggconsort object>)}, and it combines \code{geom_consort_arrow()} and
\code{geom_consort_box()}, which can also be used individually.
\code{geom_consort_line()} draws a standalone line segment with the same styling
as the diagram lines.
}
\examples{
cohorts <- trial_data \%>\%
  cohort_start("Assessed for eligibility") \%>\%
  cohort_define(
    randomized = .full \%>\% dplyr::filter(declined != 1)
  ) \%>\%
  cohort_label(randomized = "Randomized")

consort <- cohorts \%>\%
  consort_box_add("full", 0, 10, cohort_count_adorn(cohorts, .full)) \%>\%
  consort_box_add("randomized", 0, 0, cohort_count_adorn(cohorts, randomized)) \%>\%
  consort_arrow_add("full", "bottom", "randomized", "top")

library(ggplot2)
ggplot(consort) +
  geom_consort() +
  theme_consort(margin_h = 22, margin_v = 7)
}
